requirements.txt
PyYAML
python-memcached
boto
MySQL-python
PIL
oauth2
python-dateutil
flup

# Setup and administration
fabric

# Documentation
sphinx

# Database orm
SQLAlchemy
sqlalchemy-migrate

# Testing requirements
selenium
nose
paste
unittest2
mock
